requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
colorama==0.4.6
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.0
idna==3.6
iniconfig==2.0.0
multidict==6.0.4
packaging==23.2
pluggy==1.3.0
python-dotenv==1.0.0
ruff==0.1.6
yarl==1.9.3
